:tractor: Split web-chat into smaller components
