+++ cmd/web-chat/pkg/backend/util.go
@@ -1,0 +1,30 @@
+package backend
+
+import (
+    "strings"
+    "github.com/rs/zerolog"
+)
+
+// parseZerologLevel converts a string level into zerolog.Level with a safe default
+func parseZerologLevel(s string) zerolog.Level {
+    switch strings.ToLower(s) {
+    case "trace":
+        return zerolog.TraceLevel
+    case "debug":
+        return zerolog.DebugLevel
+    case "warn", "warning":
+        return zerolog.WarnLevel
+    case "error":
+        return zerolog.ErrorLevel
+    case "fatal":
+        return zerolog.FatalLevel
+    case "panic":
+        return zerolog.PanicLevel
+    case "info":
+        fallthrough
+    default:
+        return zerolog.InfoLevel
+    }
+}
+
+
